LayoutTests/http/tests/cache/disk-cache/disk-cache-range-expected.txt

Test that request with different Range header values work.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--------Testing loads from disk cache--------
response headers: {"Cache-control":"max-age=100"}
response source: Disk cache
response status: 200

response headers: {"Cache-control":"max-age=0"}
response source: Network
response status: 200

response headers: {"Cache-control":"max-age=100","Range":"bytes=2-4"}
response source: Disk cache
response status: 206

response headers: {"Cache-control":"max-age=0","Range":"bytes=2-4"}
response source: Network
response status: 206

response headers: {"Cache-control":"max-age=100","Range":"bytes=5-7"}
response source: Network
response status: 416

response headers: {"Cache-control":"max-age=0","Range":"bytes=5-7"}
response source: Network
response status: 416

--------Testing loads through memory cache (XHR behavior)--------
response headers: {"Cache-control":"max-age=100"}
response source: Memory cache
response status: 200

response headers: {"Cache-control":"max-age=0"}
response source: Network
response status: 200

response headers: {"Cache-control":"max-age=100","Range":"bytes=2-4"}
response source: Memory cache
response status: 206

response headers: {"Cache-control":"max-age=0","Range":"bytes=2-4"}
response source: Network
response status: 206

response headers: {"Cache-control":"max-age=100","Range":"bytes=5-7"}
response source: Memory cache
response status: 416

response headers: {"Cache-control":"max-age=0","Range":"bytes=5-7"}
response source: Network
response status: 416

--------Testing loads through memory cache (subresource behavior)--------
response headers: {"Cache-control":"max-age=100"}
response source: Memory cache
response status: 200

response headers: {"Cache-control":"max-age=0"}
response source: Network
response status: 200

response headers: {"Cache-control":"max-age=100","Range":"bytes=2-4"}
response source: Memory cache
response status: 206

response headers: {"Cache-control":"max-age=0","Range":"bytes=2-4"}
response source: Network
response status: 206

response headers: {"Cache-control":"max-age=100","Range":"bytes=5-7"}
response source: Memory cache
response status: 416

response headers: {"Cache-control":"max-age=0","Range":"bytes=5-7"}
response source: Network
response status: 416

PASS successfullyParsed is true

TEST COMPLETE
